requirements.txt
boto3==1.26.126
bcrypt
fastapi==0.95.1
passlib==1.7.4
pydantic==1.10.7
pydantic[email]
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn==0.22.0
pre-commit
black
pillow
